pyproject.toml
[project]
name = "koetype"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "rumps>=0.4.0",
    "pynput>=1.7.6",
    "sounddevice>=0.4.6",
    "silero-vad>=5.1",
    "transformers>=4.40.0",
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "anthropic>=0.25.0",
    "pyperclip>=1.8.2",
    "accelerate>=0.24.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "boto3>=1.34.0",
    "pyobjc-framework-ApplicationServices>=9.0",
]
